+++ src/Handlers/RequestHandler.ts
@@ -72,7 +72,10 @@ export default async (
   };
 
   response.setHeader("Content-Type", "application/json");
-  response.setHeader("x-powered-by", "Axe API");
+
+  if (api.config.disableXPoweredByHeader === false) {
+    response.setHeader("x-powered-by", "Axe API");
+  }
 
   for (const phase of match.phases) {
     // If there is an non-async phase, it should be an Event function