+++ tests/integrations/scenarios/app/v1/init.ts
@@ -1,5 +1,24 @@
 import cors from "cors";
-import { App, AxeRequest, AxeResponse } from "axe-api";
+import { App, AxeRequest, AxeResponse, createRateLimitter } from "axe-api";
+import { IncomingMessage, ServerResponse } from "http";
+
+const customRateLimitMiddleware = (
+  req: IncomingMessage,
+  res: ServerResponse,
+  next: any,
+) =>
+  createRateLimitter(
+    {
+      name: "Custom-Limitter",
+      clientKey: "test-key",
+      setResponseHeaders: true,
+      maxRequests: 1,
+      windowInSeconds: 5,
+    },
+    req,
+    res,
+    next,
+  );
 
 const onBeforeInit = async (app: App) => {
   app.use(
@@ -10,6 +29,14 @@ const onBeforeInit = async (app: App) => {
   app.get("/health/before", async (req: AxeRequest, res: AxeResponse) => {
     res.json({ health: true });
   });
+
+  app.get(
+    "/api/v1/custom-rate-limit",
+    customRateLimitMiddleware,
+    async (req: AxeRequest, res: AxeResponse) => {
+      res.json({ health: true });
+    },
+  );
 };
 
 const onAfterInit = async (app: App) => {